Relationships in migration create command

+++ src/Controllers/RelationShips.php
@@ -4,26 +4,100 @@ namespace Tgozo\LaravelCodegen\Controllers;
 
 class RelationShips
 {
-    const RELATIONSHIPS = [
-        'belongsTo',
-        'belongsToMany',
-        'hasMany',
-        'hasManyThrough',
-        'hasOne',
-        'hasOneOrMany',
-        'hasOneThrough',
-        'morphMany',
-        'morphOne',
-        'morphOneOrMany',
-        'morphPivot',
-        'morphTo',
-        'morphToMany',
+
+    const RELATIONSHIP_MAPPER = [
+        'belongsto' => [
+            'returnName' => 'BelongsTo',
+            'methodName' => 'belongsTo',
+            'plural' => false
+        ],
+        'belongstomany' => [
+            'returnName' => 'BelongsToMany',
+            'methodName' => 'belongsToMany',
+            'plural' => true
+        ],
+        'hasmany' => [
+            'returnName' => 'HasMany',
+            'methodName' => 'hasMany',
+            'plural' => true
+        ],
+        'hasmanythrough' => [
+            'returnName' => 'HasManyThrough',
+            'methodName' => 'hasManyThrough',
+            'plural' => true
+        ],
+        'hasone' => [
+            'returnName' => 'HasOne',
+            'methodName' => 'hasOne',
+            'plural' => false
+        ],
+        'hasonethrough' => [
+            'returnName' => 'HasOneThrough',
+            'methodName' => 'hasOneThrough',
+            'plural' => false
+        ],
+        'morphmany' => [
+            'returnName' => 'MorphMany',
+            'methodName' => 'morphMany',
+            'plural' => true
+        ],
+        'morphone' => [
+            'returnName' => 'MorphOne',
+            'methodName' => 'morphOne',
+            'plural' => false
+        ],
+        'morphto' => [
+            'returnName' => 'MorphTo',
+            'methodName' => 'morphTo',
+            'plural' => false
+        ],
+        'morphtomany' => [
+            'returnName' => 'MorphToMany',
+            'methodName' => 'morphToMany',
+            'plural' => true
+        ],
     ];
 
     const ARGUMENTS_REQUIRED = [
-        'belongsTo' => [
+        'belongsto' => [
             'min' => 1,
             'max' => 4
+        ],
+        'belongstomany' => [
+            'min' => 1,
+            'max' => 7
+        ],
+        'hasmany' => [
+            'min' => 1,
+            'max' => 3
+        ],
+        'hasmanythrough' => [
+            'min' => 2,
+            'max' => 6
+        ],
+        'hasone' => [
+            'min' => 1,
+            'max' => 3
+        ],
+        'hasonethrough' => [
+            'min' => 2,
+            'max' => 6
+        ],
+        'morphmany' => [
+            'min' => 2,
+            'max' => 5
+        ],
+        'morphone' => [
+            'min' => 2,
+            'max' => 5
+        ],
+        'morphto' => [
+            'min' => 0,
+            'max' => 4
+        ],
+        'morphtomany' => [
+            'min' => 2,
+            'max' => 9
         ]
     ];
 
@@ -36,13 +110,24 @@ class RelationShips
         $str = '';
         $relationType = $relationship['type'];
         $relationModel = $relationship['model'];
-        $relationName = $this->package->singularize($this->package->str_to_lower($relationModel));
+
+        $action = "singularize";
+        if (self::RELATIONSHIP_MAPPER[$relationType]['plural']){
+            $action = "pluralize";
+        }
+        $relationName = $this->package->{$action}($this->package->str_to_lower($relationModel));
+
+        $methodName = self::RELATIONSHIP_MAPPER[$relationType]['methodName'];
+        $returnName = self::RELATIONSHIP_MAPPER[$relationType]['returnName'];
+
         foreach ($relationship['parameters'] as $parameter) {
             $str = ", '{$parameter}'";
         }
 
-        $relationString = "\n\tpublic function {$relationName}(): $relationType" . "\n\t{\n\t\t";
-        $relationString .= "return \$this->{$relationType}({$relationModel}::class{$str});" . "\n\t}\n";
+        $this->package->registerNamespace("Illuminate\Database\Eloquent\Relations\\$returnName");
+
+        $relationString = "\n\tpublic function {$relationName}(): $returnName" . "\n\t{\n\t\t";
+        $relationString .= "return \$this->{$methodName}({$relationModel}::class{$str});" . "\n\t}\n";
 
         return $relationString;
     }
@@ -57,16 +142,25 @@ class RelationShips
         return trim($str);
     }
 
-    public function generateRelationships()
+    public function generateRelationships(): void
     {
         $str = $this->getRelationsString();
         $file_to_use = app_path('Models') . "/{$this->modelName}.php";
+
+        if (!file_exists($file_to_use)){
+            $this->package->warn("Relationships not generated because the file {$file_to_use} doesn't exist");
+            return;
+        }
+
         $file_contents = file_get_contents($file_to_use);
         $class_pos = strpos($file_contents, "class {$this->modelName} extends Model");
         $open_brace_pos = strpos($file_contents, '{', $class_pos);
         $close_brace_pos = $this->package->findClosingBrace($file_contents, $open_brace_pos + 1);
 
         $new_file_contents = substr_replace($file_contents, "\t$str\n", $close_brace_pos, 0);
+
+        $new_file_contents = $this->package->addNamespaces($new_file_contents, false);
+
         file_put_contents($file_to_use, $new_file_contents);
 
         $this->package->info("Relationships generated successfully in {$file_to_use}.");
